# modules/apps/app-builder/app-builder-rest-impl/rest-openapi.yaml
components:
    schemas:
        App:
            properties:
                appDeployments:
                    items:
                        $ref: "#/components/schemas/AppDeployment"
                    type: array
                dataDefinitionId:
                    format: int64
                    type: integer
                dataLayoutId:
                    format: int64
                    type: integer
                dataListViewId:
                    format: int64
                    type: integer
                dateCreated:
                    format: date-time
                    type: string
                dateModified:
                    format: date-time
                    type: string
                id:
                    format: int64
                    type: integer
                name:
                    additionalProperties:
                        type: object
                    type: object
                siteId:
                    format: int64
                    type: integer
                status:
                    type: string
                userId:
                    format: int64
                    type: integer
            type: object
        AppDeployment:
            properties:
                settings:
                    additionalProperties:
                        type: object
                    type: object
                type:
                    type: string
            type: object
        DeploymentAction:
            enum: [deploy, undeploy]
            type: string
info:
    description: ""
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: "App Builder"
    version: v1.0
openapi: 3.0.1
paths:
    "/apps/{appId}":
        delete:
            operationId: deleteApp
            parameters:
                - in: path
                  name: appId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                204:
                    content:
                        application/json: {}
                    description: ""
            tags: ["App"]
        get:
            operationId: getApp
            parameters:
                - in: path
                  name: appId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/App"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/App"
                    description: ""
            tags: ["App"]
        put:
            operationId: putApp
            parameters:
                - in: path
                  name: appId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/App"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/App"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/App"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/App"
                    description: ""
            tags: ["App"]
    "/apps/{appId}/deployment":
        put:
            operationId: putAppDeployment
            parameters:
                - in: path
                  name: appId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: deploymentAction
                  required: true
                  schema:
                      $ref: "#/components/schemas/DeploymentAction"
            responses:
                200:
                    description: "OK"
            tags: ["App"]
    "/data-definitions/{dataDefinitionId}/apps":
        get:
            operationId: getDataDefinitionAppsPage
            parameters:
                - in: path
                  name: dataDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: keywords
                  schema:
                      type: string
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
                - in: query
                  name: sort
                  required: false
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/App"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/App"
                                type: array
                    description: ""
            tags: ["App"]
        post:
            operationId: postDataDefinitionApp
            parameters:
                - in: path
                  name: dataDefinitionId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/App"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/App"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/App"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/App"
                    description: ""
            tags: ["App"]
    "/sites/{siteId}/apps":
        get:
            operationId: getSiteAppsPage
            parameters:
                - in: path
                  name: siteId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: keywords
                  schema:
                      type: string
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
                - in: query
                  name: sort
                  required: false
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/App"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/App"
                                type: array
                    description: ""
            tags: ["App"]